# docs/04-openapi.yaml
openapi: 3.1.0
info:
  title: SaCMS Public API
  version: 0.2.0
  description: |
    Stable public subset of the SaCMS headless content API.
    All requests require `Authorization: Bearer <token>` header.
    Read-only tokens can only access PUBLISHED content. Collection mutation is exposed through GraphQL, not REST.
  contact:
    name: SaCMS
servers:
  - url: "{baseUrl}"
    variables:
      baseUrl:
        default: http://localhost:3000
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API token prefixed with `cf_`
...
  parameters:
    TenantSlug:
      name: tenant
      in: path
      required: true
      schema:
        type: string
      description: Tenant slug identifier
...
paths:
...
  /api/public/{tenant}/graphql:
    post:
      summary: Execute GraphQL query or mutation
      description: |
        Mutations require `full-access` API token.
        Schema is dynamically generated from tenant's content types.
      operationId: graphql
      tags: [GraphQL]
      parameters:
        - $ref: "#/components/parameters/TenantSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  example: "{ articles { data { id title } } }"
                variables:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: GraphQL response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        message:
                          type: string
        "403":
          description: Mutation attempted with read-only token
...